RTC 125267:  Correct ambiguity in -b switch.
The 'modulemaker' help message heretofore suggested that the '-b' command-line
switch would take a string argument -- presumably something like
'ExtUtils::ModuleMaker' or 'Module::Build'. However, if you provided *any*
argument, you ended up with a generated Build.PL.

The problem was that inside ExtUtils::ModuleMaker::Opts's invocation of
'getopts', '-b' was defined as a flag rather than as an option which would
take an argument.  That meant that providing the '-b' flag would, in
ExtUtils::ModuleMaker::Interactive set BUILD_SYSTEM to 'Module::Build'
regardless of what you (mistakenly) provided as a value for '-b'.

Here we correct the help message to prevent the user from thinking that
'-b' takes a value and instruct the user that it directly selects
Module::Build as the build system.  This is probably what a sophisticated user
of the 'modulemaker' utility would want.  We change code in Interactive.pm and
Opts.pm to implement this change.  If the user does not suppress the
INTERACTIVE mode, the user can still call up the Build Menu and change to any
of the other build system options provided there.

https://rt.cpan.org/Ticket/Display.html?id=125267
Signed-off-by: James E Keenan <jkeenan@cpan.org>