Move MockHomeDir.pm to underneath lib.
The functions exported by ExtUtils::ModuleMaker::Auxiliary are intended to be
used in testing this distribution -- and all distributions which subclass it,
such as ExtUtils::ModuleMaker::PBP.  Auxiliary.pm uses functions found in
MockHomeDir.pm.  But unless MockHomeDir.pm is underneath lib/ExtUtils/, it
won't be indexed by CPAN and won't be found by the test suites of subclasses.
Hence, move it to underneath lib/.

Add $VERSION to MockHomeDir.pm.